Added files via upload
Essentially, this code looks at the randomness of the random number seed generator provided by the random number generator in the library, called random. Technically, computers don't have a good history in relation to their ability to generate random numbers as the code they use for this is determined, and therefore some level of prediction can be possible. For fun, I thought I'd look at generating paired dice rolls, just like in a craps game. Probability theory shows that in 2-pair dice throws that number 7 should be most frequently received (2 & 5, 3 & 4, 1 & 6) with a probability of 16.7%. 2 and 12 should have least probability as they are the rarest event that can happen (as only double 6 and double 1 can yield this result). So, if you use the random generator to throw the pairs, say, a million times - then when a histogram of results is presented, it should highlight precise probabilities via a consistent and smooth histogram. In the code presented, I have set the number of throws